bdei {phylopomp}	R Documentation

Linear birth-death with exposed and infectious classes

Description

Two-deme linear birth-death process with exposed (incubating) and infectious lineages. Transmission creates exposed offspring from infectious donors; exposed lineages progress to infectious. Only infectious lineages are subject to sampling and sampling is destructive. Identical to the BDEI model of Voznica et al. (2022, https://doi.org/10.1038/s41467-022-31511-0), albeit with a different parameterization.

Usage

runBDEI(
  time,
  t0 = 0,
  sigma = 1/7,
  lambda = 7/9,
  mu = 1/6,
  chi = 1/6,
  pop = 1,
  E0 = 0,
  I0 = 1
)

continueBDEI(object, time, sigma = NA, lambda = NA, mu = NA, chi = NA)

Arguments

time
end timepoint of simulation

t0
beginning timepoint of simulation

sigma
per capita progression rate (exposed to infectious)

lambda
per capita reproduction rate

mu
per capita removal rate (of infectious lineages)

chi
(destructive) sampling rate

pop
initial population size

E0
initial fraction of exposed lineages

I0
initial fraction of infectious lineages

object
a previously computed simulation

Value

runBDEI and continueBDEI return objects of class ‘gpsim’ with ‘model’ attribute “BDEI”.

References

J. Voznica, A. Zhukova, V. Boskova, E. Saulnier, F. Lemoine, M. Moslonka-Lefebvre, O. Gascuel. Deep learning from phylogenies to uncover the epidemiological dynamics of outbreaks. Nature Communications 13, 3896, 2022. doi:10.1038/s41467-022-31511-0.

See Also

More example genealogy processes: bdss, lbdp, mers, moran, s2i2r2, seir, si2r, siir, simulate(), sir, strains, twospecies, twoundead

Examples

library(phylopomp)

runBDEI(time=10, pop=2, E0=0.5, I0=0.5) |>
  plot()

[Package phylopomp version 0.19.4.1 Index]