diagram {phylopomp}	R Documentation

Genealogy process diagram

Description

Produces a diagram of the genealogy process state.

Usage

diagram(
  object,
  prune = TRUE,
  obscure = TRUE,
  m = NULL,
  n = NULL,
  ...,
  digits = 1
)

## S3 method for class 'gpdiag'
print(x, newpage = is.null(vp), vp = NULL, ...)

Arguments

object
gpsim object.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

m
width of plotting window, in nodes. By default, the nodes will be adjusted in width to fit the window.

n
height of the pockets, in balls. By default, the balls will be adjusted in size to fit the space available.

...
other arguments, ignored.

digits
non-negative integer; number of decimal digits to print in the node time

x
An R object.

newpage
draw new empty page first?

vp
viewport to draw plot in

Value

A grid graphics object (grob), invisibly.

Examples

runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot(x,points=TRUE,prune=FALSE)
plot_grid(plotlist=list(plot(x,points=TRUE)[[1]],diagram(x)),
  ncol=1,rel_heights=c(4,1))

[Package phylopomp version 0.6.0.0 Index]