getInfo {phylopomp}	R Documentation

getInfo

Description

Retrieve information from genealogy process simulation

Usage

getInfo(
  object,
  prune = TRUE,
  obscure = TRUE,
  t0 = FALSE,
  time = FALSE,
  description = FALSE,
  structure = FALSE,
  yaml = FALSE,
  lineages = FALSE,
  tree = FALSE,
  compact = TRUE
)

Arguments

object
gpsim object.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

t0
logical; return the zero-time?

time
logical; return the current time?

description
logical; return the description?

structure
logical; return the structure in R list format?

yaml
logical; return the structure in YAML format?

lineages
logical; return the lineage-count function?

tree
logical; return the tree?

compact
logical; return the tree in compact representation?

Value

A list containing the requested elements, including any or all of:

t0

the initial time

time

the current time

tree

the genealogical tree, in Newick format

description

a human readable description of the state of the genealogy process

yaml

the state of the genealogy process in YAML format

structure

the state of the genealogy process in R list format

lineages

a tibble containing the lineage count function through time

Examples

simulate("SIIR",time=3,psi1=1,psi2=0) |>
  simulate(Beta1=2,gamma=2,time=10,psi1=10,psi2=1) |>
  plot()

runSIIR(Beta1=10,Beta2=8,
  S0=200,I1_0=10,I2_0=8,R0=0,time=0,t0=-1) |>
  simulate(psi1=10,time=2) |>
  plot(points=TRUE,obscure=FALSE)

simulate("SIIR",Beta1=2,Beta2=50,gamma=1,psi1=2,
  S0=300,I1_0=20,I2_0=2,time=5) |>
  lineages() |>
  plot()

[Package phylopomp version 0.1.0.0 Index]