tableauGrob {phylopomp}	R Documentation

Diagramming internals

Description

Facilities to produce diagrammatic representations of genealogy process states.

Usage

tableauGrob(data, n, vp = NULL)

playerGrob(name, ballA, ballAcol, ballB, ballBcol, slate, vp = NULL)

resizingTextGrob(y, ..., vp = NULL)

## S3 method for class 'resizingTextGrob'
drawDetails(x, recording = TRUE)

## S3 method for class 'resizingTextGrob'
preDrawDetails(x)

## S3 method for class 'resizingTextGrob'
postDrawDetails(x)

ballGrob(y, label, color, ..., vp = NULL)

## S3 method for class 'ballGrob'
drawDetails(x, recording = TRUE)

## S3 method for class 'ballGrob'
preDrawDetails(x)

## S3 method for class 'ballGrob'
postDrawDetails(x)

Arguments

data
illustration vector

n
length of longest tableau

vp
a viewport object (or NULL).

name
player name

ballA, ballB
ball names

ballAcol, ballBcol
ball colors

slate
time

y
height

...
arguments to be passed to textGrob.

x
An R object.

recording
A logical value indicating whether a grob is being added to the display list or redrawn from the display list.

label
ball name

color
ball color

Details

Code for the resizing text adapted from a blog post by Mark Heckmann (https://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/).

[Package phylopomp version 0.0.23.1 Index]