lbdp {phylopomp}	R Documentation

Linear birth-death process.

Description

Simulation and inference based on linear birth-death-sampling processes.

Usage

playLBDP(
  data = NULL,
  lambda,
  mu,
  psi,
  n0 = 1,
  t0 = 0,
  times,
  tree = FALSE,
  ill = FALSE
)

lbdp_exact(data, lambda, mu, psi)

lbdp_pomp(data, lambda, mu, psi, n0 = 1, t0 = 0)

Arguments

data
data frame containing the genealogy event times and event codes.

lambda
birth rate

mu
death rate

psi
sampling rate

n0
initial population size

t0
initial time

times
times at which output is requested.

tree
logical; represent the genealogical tree in Newick format?

ill
logical; return an illustration?

Details

playLBDP is a simulator of the genealogy process induced by a linear birth-death-sampling process.

lbdp_exact gives the exact likelihood of a linear birth-death process, conditioned on n0=0 (Stadler, 2010, Thm 3.5). The derivation is also given in comments in the code.

The data argument should in the format returned by newick2df.

lbdp_pomp constructs a pomp object containing a given set of data and a linear birth-death-sampling process.

It is assumed that data is in the format returned by newick2df.

Value

playLBDP returns an object of class gpsim. Available methods for such objects include getInfo and plot.

lbdp_exact returns the log likelihood of the genealogy. Note that the time since the most recent sample is informative.

References

T. Stadler. Sampling-through-time in birth-death trees. Journal of Theoretical Biology 267, 396–404, 2010.

See Also

Other Genealogy processes: leventhal, moran, sirs, sirws

Examples

playLBDP(lambda=1,mu=0.5,psi=1,n0=3,times=seq(0,4,by=0.2),tree=TRUE) -> x
plot(x,points=TRUE)

y <- getInfo(x,prune=FALSE)
plot(y,points=TRUE)

library(ggplot2)
y$lineages |>
  ggplot(aes(x=time,y=lineages))+
  geom_step()+
  geom_vline(xintercept=y$etimes,alpha=0.1)

y$cumhaz |>
  ggplot(aes(x=exp(-Lambda)))+
  stat_ecdf()+
  geom_abline(slope=1)

playLBDP(lambda=2,mu=1,psi=0.5,n0=1,times=5) |>
  getInfo() -> x
plot(getInfo(x,compact=TRUE),points=TRUE)

library(pomp)
x$tree |>
  newick2df(time=5) |>
  lbdp_pomp(psi=0.5,lambda=2,mu=1,n0=1) |>
  pfilter(Np=2000) |>
  logLik()

x$tree |>
  newick2df(time=5) |>
  lbdp_exact(psi=0.5,lambda=2,mu=1)

[Package phylopomp version 0.0.23.1 Index]