leventhal {phylopomp}	R Documentation

Leventhal (2014) SI model

Description

Simulation and inference for Leventhal's model

Usage

leventhal_pomp(data, beta, gamma, psi, N, I0 = 1, t0 = 0)

Arguments

data
optional data frame; output from playLeventhal.

beta
contact rate

gamma
recovery rate

psi
sampling rate

N
population size

I0
initial number of infections

t0
initial time

Details

leventhal_pomp constructs a pomp object containing a given set of data and a Leventhal-type SIS infection.

It is assumed that data is in the format returned by newick2df.

See Also

Other Genealogy processes: lbdp, moran, sirs, sirws

[Package phylopomp version 0.0.23.1 Index]