newick {phylopomp}	R Documentation

Newick output

Description

Extract a Newick-format description of a genealogy.

Usage

newick(object, prune = TRUE, obscure = TRUE, extended = TRUE)

Arguments

object
gpsim object.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

extended
logical; if TRUE, an extended-Newick format is used. See Details.

Details

In the extended-Newick format, metadata tags of the form [&&PhyloPOMP ...] are inserted. These contain information regarding node-type and deme. See below for more information.

If extended = FALSE, then the string returned is in ordinary Newick format. Deme and node-type information is discarded. Every sample becomes a tip (inline samples lie on the ends of branches of zero length). Note that extended = FALSE implies both prune = TRUE and obscure = TRUE.

Value

A string in (possibly extended) Newick format.

Metadata tags

Metadata tags are of the form

[&&PhyloPOMP type={node|sample|extant|root|migration|branch} deme=<integer>].

Types ‘branch’, ‘node’, ‘migration’, and ‘root’ are equivalent on input to parse_newick; ‘type=node’ is used in output of newick. Note that 0 is reserved for the “undeme”, i.e., unspecified or unknown deme.

Examples

simulate("SIIR",time=0.3,sigma12=1) -> x

x |> plot(prune=FALSE,obscure=FALSE)

x |> newick(prune=FALSE,obscure=FALSE)

x |> newick()

x |> newick(extended=FALSE)

x |>
  newick() |>
  parse_newick(t0=0)

x |>
  newick(prune=FALSE,obscure=FALSE) |>
  parse_newick(t0=0)

x |>
  newick(extended=FALSE) |>
  parse_newick(t0=0)

"" |> parse_newick() |> newick()

";" |> parse_newick() |> newick()

"((:1,:1):1,:1):1;" |> parse_newick() |> newick()

[Package phylopomp version 0.19.3.1 Index]