newick2df {phylopomp}	R Documentation

Convert a tree in Newick format to data frame

Description

Convert a genealogical tree in Newick format to a data frame suitable for use with pomp.

Usage

newick2df(tree, time = NA, root_time = 0)

Arguments

tree
tree data in Newick format.

time
time of the genealogy.

root_time
time of the root.

Details

If time is furnished, it is assumed that the absence of samples between the latest leaf and time is informative.

Invisible nodes (labeled 'X_' for any X) are dropped.

Value

A data frame suitable for use as pomp input, containing three columns:

time

numeric; time of the genealogy event.

lineages

integer; the value of the lineage-count function at the specified time. Note that this function is right-continuous with left limits, and constant on the inter-event intervals.

code

integer; a code describing the nature of the event. 1 indicates a coalescence; 0 indicates a dead sample; -1 indicates a live sample; 2 indicates the root.

Examples

runSIR(Beta=2,gamma=1,psi=2,S0=100,I0=2,R0=0,time=5,t0=0) |>
  getInfo(tree=TRUE) |>
  getElement("tree") |>
  newick2df()
  

[Package phylopomp version 0.0.999.9 Index]