parse_newick {phylopomp}	R Documentation

parse a Newick string

Description

Parses a Newick description and returns a binary version of the genealogy.

Usage

parse_newick(x, t0 = 0, time = NA)

Arguments

x
character; the Newick description. See Details for specifics.

t0
numeric; the root time.

time
numeric; the current or final time.

Details

parse_newick parses a string containing information in a (possibly extended) Newick format and returns a phylopomp genealogy.

The extension allows metadata (enclosed by square brackets). It ignores metadata unless they are specifically flagged as PhyloPOMP metadata, i.e.: [&&PhyloPOMP ...]. PhyloPOMP metadata can be used to encode the node type (i.e., type=sample, type=extant, type=node) and/or the deme (given as an integer, e.g., deme=1). See below for more information.

Value

An object of class “gpgen”.

Metadata tags

Metadata tags are of the form

[&&PhyloPOMP type={node|sample|extant|root|migration|branch} deme=<integer>].

Types ‘branch’, ‘node’, ‘migration’, and ‘root’ are equivalent on input to parse_newick; ‘type=node’ is used in output of newick. Note that 0 is reserved for the “undeme”, i.e., unspecified or unknown deme.

Examples

simulate("SIIR",time=0.3,sigma12=1) -> x

x |> plot(prune=FALSE,obscure=FALSE)

x |> newick(prune=FALSE,obscure=FALSE)

x |> newick()

x |> newick(extended=FALSE)

x |>
  newick() |>
  parse_newick(t0=0)

x |>
  newick(prune=FALSE,obscure=FALSE) |>
  parse_newick(t0=0)

x |>
  newick(extended=FALSE) |>
  parse_newick(t0=0)

"" |> parse_newick() |> newick()

";" |> parse_newick() |> newick()

"((:1,:1):1,:1):1;" |> parse_newick() |> newick()

[Package phylopomp version 0.19.5.2 Index]