plot {phylopomp}	R Documentation

Fancy tree plotter

Description

Plots a genealogical tree.

Usage

## S3 method for class 'gpgen'
plot(
  x,
  ...,
  prune = TRUE,
  obscure = TRUE,
  points = FALSE,
  ladderize = TRUE,
  palette = scales::hue_pal(l = 30, h = c(220, 580))
)

Arguments

x
object of class ‘gpgen’

...
plot passes extra arguments to theme.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

points
logical; show nodes and tips?

ladderize
logical; ladderize?

palette
color palette for indicating demes. This can be furnished either as a function or a vector of colors. If this is a function, it should take a single integer argument, the number of colors required. If it is a vector, it should have at least as many elements as there are demes in the genealogy.

Details

Tree-plotting in phylopomp depends on the ggtree package.

Value

A printable ggplot object.

References

G. Yu, D. K. Smith, H. Zhu, Y. Guan, T. T. Y. Lam. ggtree: an R package for visualization and annotation of phylogenetic trees with their covariates and other associated data. Methods in Ecology and Evolution 8, 28–36, 2017. doi:10.1111/2041-210X.12628.

Examples

simulate("SEIR",Beta=2,sigma=2,gamma=1,psi=2,S0=1,I0=0.01,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSEIR(Beta=3,gamma=1,psi=2,S0=1,I0=0.2,R0=0,time=5,t0=-1) |>
  plot(points=TRUE,obscure=FALSE)

runSEIR(Beta=3,gamma=0.1,psi=0.2,S0=1,I0=0.05,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SEIR",sigma=1,omega=1,time=20,I0=0.04) |> plot(obscure=FALSE)

simulate("SEIR",sigma=1,omega=1,time=20,I0=0.04) |>
  lineages(obscure=FALSE) |>
  plot()
simulate("LBDP",time=4) |> plot(points=TRUE)

simulate("LBDP",lambda=2,mu=1,psi=3,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot()

simulate("LBDP",time=4) |>
  lineages() |>
  plot()

runS2I2R2(
  time=30,
  iota2=0.01,Beta12=0.1,
  I1_0=0,I2_0=0,
  psi1=10,psi2=10,
  omega1=0.2,
  b2=0.02,d2=0.02
) |>
  freeze(seed=847110120) -> x

plot_grid(
  x |>
    plot(
      obscure=FALSE,
      palette=c("#000000ff","#440154ff","#21908cff","#ccccccff"),
      legend.position="inside",
      legend.position.inside=c(0.7,0.7)
    ),
  x |>
    lineages(obscure=FALSE) |>
    plot(
      palette=c("#000000ff","#440154ff","#21908cff","#ccccccff"),
      legend.position="none"
    ),
  ncol=1,
  align="hv",
  axis="b"
)
## Not run: 

  library(ggplot2)
  library(phylopomp)
  times <- seq(from=0,to=8,by=0.1)[-1]

  png_files <- sprintf(
    file.path(tempdir(),"frame%05d.png"),
    seq_len(2*length(times))
  )

  pb <- utils::txtProgressBar(0,2*length(times),0,style=3)
  x <- simulate("SIIR",time=0,Beta1=5,Beta2=10,gamma=1,omega=0.5,
    psi1=0.2,psi2=0.1,sigma12=1,sigma21=1,pop=205,S_0=200,I1_0=3,I2_0=2)

  img <- 1
  for (k in seq.int(from=1,to=length(times),by=1)) {
    x <- simulate(x,time=times[k])
    ggsave(
      filename=png_files[img],
      plot=plot(
        x,
        points=FALSE, prune=FALSE, obscure=FALSE,
        ladderize=FALSE,
        palette=c("#000000","#ffcb05","#dddddd"),
        axis.line=element_line(color="white"),
        axis.ticks=element_line(color="white"),
        axis.text=element_blank(),
        plot.background=element_rect(fill=NA,color=NA),
        panel.background=element_rect(fill=NA,color=NA),
        legend.position="none"
      )+
        expand_limits(x=c(0,range(times))),
      device="png",dpi=300,
      height=2,width=3,units="in"
    )
    setTxtProgressBar(pb,img)
    img <- img+1
  }

  for (k in seq.int(from=length(times),to=1,by=-1)) {
    x <- curtail(x,time=times[k])
    ggsave(
      filename=png_files[img],
      plot=plot(
        x,
        points=FALSE, prune=FALSE, obscure=FALSE,
        ladderize=FALSE,
        palette=c("#000000","#ffcb05","#dddddd"),
        axis.line=element_line(color="white"),
        axis.ticks=element_line(color="white"),
        axis.text=element_blank(),
        plot.background=element_rect(fill=NA,color=NA),
        panel.background=element_rect(fill=NA,color=NA),
        legend.position="none"
      )+
        expand_limits(x=c(0,range(times))),
      device="png",dpi=300,
      height=2,width=3,units="in"
    )
    setTxtProgressBar(pb,img)
    img <- img+1
  }

  library(gifski)
  gif_file <- "movie1.gif"
  gifski(png_files,gif_file,delay=0.08,loop=TRUE)
  unlink(png_files)

## End(Not run)

[Package phylopomp version 0.19.6.0 Index]