s2i2r2 {phylopomp}	R Documentation

Two-host infection model with waning, immigration, and demography.

Description

The population is structured by infection progression and host species.

Usage

runS2I2R2(
  time,
  t0 = 0,
  Beta11 = 4,
  Beta12 = 0,
  Beta22 = 4,
  gamma1 = 1,
  gamma2 = 1,
  psi1 = 1,
  psi2 = 0,
  omega1 = 0,
  omega2 = 0,
  b1 = 0,
  b2 = 0,
  d1 = 0,
  d2 = 0,
  iota1 = 0,
  iota2 = 0,
  S1_0 = 100,
  S2_0 = 100,
  I1_0 = 0,
  I2_0 = 10,
  R1_0 = 0,
  R2_0 = 0
)

continueS2I2R2(
  object,
  time,
  Beta11 = NA,
  Beta12 = NA,
  Beta22 = NA,
  gamma1 = NA,
  gamma2 = NA,
  psi1 = NA,
  psi2 = NA,
  omega1 = NA,
  omega2 = NA,
  b1 = NA,
  b2 = NA,
  d1 = NA,
  d2 = NA,
  iota1 = NA,
  iota2 = NA
)

Arguments

time
final time

t0
initial time

Beta11, Beta22
transmission rates within species 1 and 2, respectively

Beta12
transmission from species 2 to species 1

gamma1, gamma2
recovery rates for species 1 and 2, respectively

psi1, psi2
per capita sampling rates

omega1, omega2
rates of waning of immunity

b1, b2
per capita birth rates

d1, d2
per capita death rates

iota1, iota2
infection importation rates

S1_0, S2_0
initial sizes of susceptible populations

I1_0, I2_0
initial sizes of infected populations

R1_0, R2_0
initial sizes of immune populations

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

Value

runS2I2R2 and continueS2I2R2 return objects of class ‘gpsim’ with ‘model’ attribute “S2I2R2”.

See Also

More example genealogy processes: lbdp, moran, seir, si2r, siir, simulate(), sir, strains, twospecies, twoundead

[Package phylopomp version 0.15.2.0 Index]