seir {phylopomp}	R Documentation

Classical susceptible-exposed-infected-recovered model

Description

The population is structured by infection progression. There is an incubation period that must pass before an infected host becomes infectious. The duration of this period is 1/\sigma. The infectious period is 1/\gamma.

Usage

runSEIR(
  time,
  t0 = 0,
  Beta = 4,
  sigma = 1,
  gamma = 1,
  psi = 1,
  chi = 0,
  omega = 0,
  pop = 100,
  S0 = 0.9,
  E0 = 0.05,
  I0 = 0.05,
  R0 = 0
)

runSEIRS(
  time,
  t0 = 0,
  Beta = 4,
  sigma = 1,
  gamma = 1,
  psi = 1,
  chi = 0,
  omega = 0,
  pop = 100,
  S0 = 0.9,
  E0 = 0.05,
  I0 = 0.05,
  R0 = 0
)

continueSEIR(
  object,
  time,
  Beta = NA,
  sigma = NA,
  gamma = NA,
  psi = NA,
  chi = NA,
  omega = NA
)

continueSEIRS(
  object,
  time,
  Beta = NA,
  sigma = NA,
  gamma = NA,
  psi = NA,
  chi = NA,
  omega = NA
)

seirs_pomp(x, Beta, sigma, gamma, psi, chi = 0, omega = 0, S0, E0, I0, R0, pop)

Arguments

time
end timepoint of simulation

t0
beginning timepoint of simulation

Beta
transmission rate

sigma
progression rate

gamma
recovery rate

psi
per capita non-destructive sampling rate

chi
per capita destructive sampling rate

omega
rate of waning of immunity

pop
host population size

S0
initial fraction of population susceptible to infection

E0
initial fraction of population in exposed class

I0
initial fraction of population in infectious class

R0
initial fraction of population immune to infection

object
a previously computed simulation

x
genealogy in phylopomp format.

Details

seirs_pomp constructs a ‘pomp’ object containing a given set of data and an SEIRS model.

Value

runSEIR and continueSEIR return objects of class ‘gpsim’ with ‘model’ attribute “SEIR”.

seirs_pomp returns a ‘pomp’ object.

References

A. A. King, Q. Lin, and E. L. Ionides. Exact phylodynamic likelihood via structured Markov genealogy processes. arXiv 2405.17032, 2024. doi:10.48550/arxiv.2405.17032.

See Also

More example genealogy processes: bdei, bdss, lbdp, mers, moran, s2i2r2, si2r, siir, simulate(), sir, strains, twospecies, twoundead

Examples

simulate("SEIR",Beta=2,sigma=2,gamma=1,psi=2,S0=1,I0=0.01,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSEIR(Beta=3,gamma=1,psi=2,S0=1,I0=0.2,R0=0,time=5,t0=-1) |>
  plot(points=TRUE,obscure=FALSE)

runSEIR(Beta=3,gamma=0.1,psi=0.2,S0=1,I0=0.05,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SEIR",sigma=1,omega=1,time=20,I0=0.04) |> plot(obscure=FALSE)

simulate("SEIR",sigma=1,omega=1,time=20,I0=0.04) |>
  lineages(obscure=FALSE) |>
  plot()

[Package phylopomp version 0.19.8.0 Index]