sirs {phylopomp}	R Documentation

SIRS with sampling simulator.

Description

Run the simulator.

Usage

playSIRS(
  data = NULL,
  Beta,
  gamma,
  psi,
  Delta,
  S0,
  I0,
  R0,
  t0 = 0,
  times,
  tree = FALSE,
  ill = FALSE
)

sirs_pomp(data, Beta, gamma, psi, Delta, S0, I0, R0, t0 = 0)

Arguments

data
data frame containing the genealogy in the format returned by newick2df.

Beta
transmission rate.

gamma
recovery rate.

psi
sampling rate.

Delta
waning rate of immunity.

S0
initial size of susceptible population.

I0
initial size of infected population.

R0
initial size of recovered population.

t0
initial time

times
times at which output is requested.

tree
logical; represent the genealogical tree in Newick format?

ill
logical; return an illustration?

Details

sirs_pomp constructs a pomp object containing a given set of data and a SIR model.

Value

A tibble with state attribute.

See Also

Other Genealogy processes: lbdp, leventhal, moran, sirws

[Package phylopomp version 0.0.23.1 Index]