sirwr {phylopomp}	R Documentation

SIR model with S, M, L segments and reassortment

Description

A single, unstructured population of hosts.

Usage

runSIRwr(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  delta = 0,
  rhoS = 0,
  rhoM = 0.1,
  rhoL = 0,
  rhoSM = 0,
  rhoSL = 0,
  rhoML = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIRwr(
  object,
  time,
  Beta = NA,
  gamma = NA,
  psi = NA,
  delta = NA,
  rhoS = NA,
  rhoM = NA,
  rhoL = NA,
  rhoSM = NA,
  rhoSL = NA,
  rhoML = NA
)

Arguments

time
final time

t0
initial time

Beta
transmission rate.

gamma
recovery rate.

psi
sampling rate.

delta
immunity waning rate

rhoS
reassortment rate of S segment

rhoM
reassortment rate of M segment

rhoL
reassortment rate of L segment

rhoSM
reassortment rate of both S and M segment

rhoSL
reassortment rate of both S and L segment

rhoML
reassortment rate of both M and L segment

S0
initial size of susceptible population.

I0
initial size of infected population.

R0
initial size of recovered population.

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

Value

An object of class ‘gpsim’ with ‘model’ attribute “SIR”.

See Also

More example genealogy processes: lbdp, moran, seir, si2r, siir, simulate(), sir

Examples

simulate("SIRwr",Beta=2,gamma=1,psi=2,
         rhoS=.5, rhoM=.1, rhoL=.3, 
         rhoSM=0, rhoSL=.2, rhoML = 0,
         S0=100,I0=10,time=5) |>
  plot(prune=TRUE, compact=TRUE, points=TRUE)

runSIRwr(Beta=3,gamma=1,psi=2,
         rhoS=0,rhoM=0.5,rhoL=0,
         rhoSM=0, rhoSL=.2, rhoML = 0,
         S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)
# 
# runSIRwr(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
# plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
#   ncol=1,rel_heights=c(4,1))
# 
# simulate("SIRwr",delta=1,time=20,I0=4) |> plot()
# simulate("SIRwr",delta=1,time=20,I0=4) |> lineages() |> plot()

[Package phylopomp version 0.0.999.2 Index]