sirws {phylopomp}	R Documentation

SIR with sampling simulator.

Description

Run the simulator.

Usage

playSIRwS(
  data = NULL,
  Beta,
  gamma,
  psi,
  S0,
  I0,
  R0,
  t0 = 0,
  times,
  tree = FALSE,
  ill = FALSE
)

sir_pomp(data, Beta, gamma, psi, S0, I0, R0, t0 = 0)

Arguments

data
data frame containing the genealogy in the format returned by newick2df.

Beta
transmission rate.

gamma
recovery rate.

psi
sampling rate.

S0
initial size of susceptible population.

I0
initial size of infected population.

R0
initial size of recovered population.

t0
initial time

times
times at which output is requested.

tree
logical; represent the genealogical tree in Newick format?

ill
logical; return an illustration?

Details

sir_pomp constructs a pomp object containing a given set of data and a SIR model.

Value

A tibble with state attribute.

See Also

Other Genealogy processes: lbdp, leventhal, moran, sirs

Examples

playSIRwS(Beta=2,gamma=1,psi=2,S0=1000,I0=5,R0=0,times=0:5,t0=0,tree=TRUE) -> x
playSIRwS(x,times=6:10,psi=1,tree=TRUE) -> x
plot(x)

playSIRwS(Beta=3,gamma=1,psi=2,S0=10,I0=5,R0=0,times=0:5,t0=-1,tree=TRUE) -> x
plot(x,points=TRUE)

y <- getInfo(x)
plot(y,points=TRUE)

library(ggplot2)
y$lineages |>
  ggplot(aes(x=time,y=lineages))+
  geom_step()

[Package phylopomp version 0.0.23.1 Index]