strains {phylopomp}	R Documentation

Three strains compete for a single susceptible pool.

Description

The three demes are three distinct pathogen strains that compete for susceptibles. Sampling is assumed destructive, i.e., sampled individuals are removed from the infectious pool.

Usage

runStrains(
  time,
  t0 = 0,
  Beta1 = 5,
  Beta2 = 5,
  Beta3 = 5,
  gamma = 1,
  chi1 = 1,
  chi2 = 0,
  chi3 = 0,
  pop = 1e+05,
  S_0 = 0.8,
  I1_0 = 0.01,
  I2_0 = 0.01,
  I3_0 = 0.01,
  R_0 = 0.2
)

continueStrains(
  object,
  time,
  Beta1 = NA,
  Beta2 = NA,
  Beta3 = NA,
  gamma = NA,
  chi1 = NA,
  chi2 = NA,
  chi3 = NA
)

strains_pomp(
  x,
  Beta1,
  Beta2,
  Beta3,
  gamma,
  chi1,
  chi2,
  chi3,
  pop,
  S_0,
  I1_0,
  I2_0,
  I3_0,
  R_0
)

Arguments

time
end timepoint of simulation

t0
beginning timepoint of simulation

Beta1
transmission rate for strain 1

Beta2
transmission rate for strain 2

Beta3
transmission rate for strain 3

gamma
recovery rate

chi1
(destructive) sampling rate for strain 1

chi2
(destructive) sampling rate for strain 2

chi3
(destructive) sampling rate for strain 3

pop
population size

S_0
initial susceptible fraction

I1_0
initial fraction of population infected by strain 1

I2_0
initial fraction of population infected by strain 2

I3_0
initial fraction of population infected by strain 2

R_0
initial fraction of population immune

object
a previously computed simulation

x
genealogy in phylopomp format (i.e., an object that inherits from ‘gpgen’).

Details

strains_pomp constructs a ‘pomp’ object containing a given set of data and the Strains model.

Value

runStrains and continueStrains return objects of class ‘gpsim’ with ‘model’ attribute “Strains”.

strains_pomp returns a ‘pomp’ object.

See Also

More example genealogy processes: lbdp, mers, moran, s2i2r2, seir, si2r, siir, simulate(), sir, twospecies, twoundead

[Package phylopomp version 0.19.0.0 Index]