twoundead {phylopomp}	R Documentation

Two-host infection model with waning, immigration, demography, and spillover. Hosts are culled upon sampling with a given probability. This is identical to the TwoSpecies model with the exception that dead lineages are not pruned. Instead, they become *ghosts*.

Description

The population is structured by infection progression and host species.

Usage

runTwoUndead(
  time,
  t0 = 0,
  Beta11 = 4,
  Beta12 = 0,
  Beta21 = 0,
  Beta22 = 4,
  gamma1 = 1,
  gamma2 = 1,
  psi1 = 1,
  psi2 = 0,
  c1 = 1,
  c2 = 1,
  omega1 = 0,
  omega2 = 0,
  b1 = 0,
  b2 = 0,
  d1 = 0,
  d2 = 0,
  iota1 = 0,
  iota2 = 0,
  S1_0 = 100,
  S2_0 = 100,
  I1_0 = 0,
  I2_0 = 10,
  R1_0 = 0,
  R2_0 = 0
)

continueTwoUndead(
  object,
  time,
  Beta11 = NA,
  Beta12 = NA,
  Beta21 = NA,
  Beta22 = NA,
  gamma1 = NA,
  gamma2 = NA,
  psi1 = NA,
  psi2 = NA,
  c1 = NA,
  c2 = NA,
  omega1 = NA,
  omega2 = NA,
  b1 = NA,
  b2 = NA,
  d1 = NA,
  d2 = NA,
  iota1 = NA,
  iota2 = NA
)

Arguments

time
final time

t0
initial time

Beta11
transmission rate within species 1

Beta12
transmission from species 2 to species 1

Beta21
transmission from species 1 to species 2

Beta22
transmission rate within species 2

gamma1
species 1 recovery rate

gamma2
species 2 recovery rate

psi1
per capita sampling rate for species 1

psi2
per capita sampling rate for species 2

c1
probability that a sampled (positive) host of species 1 is culled

c2
probability that a sampled (positive) host of species 2 is culled

omega1
rate of waning of immunity for species 1

omega2
rate of waning of immunity for species 2

b1
per capita birth rate for species 1

b2
per capita birth rate for species 2

d1
per capita death rate for species 1

d2
per capita death rate for species 2

iota1
imported infections for species 1

iota2
imported infections for species 2

S1_0
initial size of species 1 susceptible population

S2_0
initial size of species 2 susceptible population

I1_0
initial size of species 1 infected population

I2_0
initial size of species 2 infected population

R1_0
initial size of species 1 immune population

R2_0
initial size of species 2 immune population

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

Value

runTwoUndead and continueTwoUndead return objects of class ‘gpsim’ with ‘model’ attribute “TwoUndead”.

See Also

More example genealogy processes: lbdp, moran, s2i2r2, seir, si2r, siir, simulate(), sir, strains, twospecies

[Package phylopomp version 0.15.0.0 Index]