as.data.frame {phylopomp}	R Documentation

Coerce to a Data Frame

Description

Functions to coerce an object to a data frame.

Usage

## S3 method for class 'gplin'
as.data.frame(x, ...)

Arguments

x
any R object.

...
additional arguments to be passed to or from methods.

Details

An object of class ‘gplin’ is coerced to a data frame by means of as.data.frame.

[Package phylopomp version 0.0.999.2 Index]