getInfo {phylopomp}	R Documentation

getInfo

Description

Retrieve information from genealogy process simulation

Usage

getInfo(
  object,
  prune = TRUE,
  obscure = TRUE,
  t0 = FALSE,
  time = FALSE,
  description = FALSE,
  structure = FALSE,
  yaml = FALSE,
  ndeme = FALSE,
  lineages = FALSE,
  newick = FALSE,
  nsample = FALSE,
  genealogy = FALSE,
  gendat = FALSE
)

Arguments

object
gpsim object.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

t0
logical; return the zero-time?

time
logical; return the current time?

description
logical; return the description?

structure
logical; return the structure in R list format?

yaml
logical; return the structure in YAML format?

ndeme
logical; return the number of demes?

lineages
logical; return the lineage-count function?

newick
logical; return a Newick-format description of the tree?

nsample
logical; return the number of samples?

genealogy
logical; return the lineage-traced genealogy?

gendat
logical; return the data-frame format?

Value

A list containing the requested elements, including any or all of:

t0

the initial time (a numeric scalar)

time

the final time (a numeric scalar)

ndeme

the number of demes (an integer)

nsample

the number of samples (an integer)

newick

the genealogical tree, in Newick format

description

a human readable description of the state of the genealogy process

yaml

the state of the genealogy process in YAML format

structure

the state of the genealogy process in R list format

lineages

a tibble containing the lineage count function through time

gendat

a tibble containing the (obscured) genealogy in a data-frame format

genealogy

the lineage-traced genealogy (as a raw vector)

Examples

simulate("SIIR",time=3,psi1=1,psi2=0) |>
  simulate(Beta1=2,gamma=2,time=10,psi1=10,psi2=1) |>
  plot()

runSIIR(Beta1=10,Beta2=8,
  S0=200,I1_0=10,I2_0=8,R0=0,time=0,t0=-1) |>
  simulate(psi1=10,time=2) |>
  plot(points=TRUE,obscure=FALSE)

simulate("SIIR",Beta1=2,Beta2=50,gamma=1,psi1=2,
  S0=300,I1_0=20,I2_0=2,time=5) |>
  lineages() |>
  plot()

[Package phylopomp version 0.14.5.0 Index]