genealogy diagram internals {phylopomp}	R Documentation

Diagramming internals

Description

Facilities to produce diagrammatic representations of genealogy process states.

Usage

genealogyGrob(object, m = NULL, n = NULL, vp = NULL, ...)

nodeGrob(object, digits = 1, n = NULL, vp = NULL)

pocketGrob(object, n = NULL, vp = NULL)

ballGrob(object, vp = NULL)

resizingTextGrob(..., vp = NULL)

## S3 method for class 'resizingTextGrob'
drawDetails(x, recording = TRUE)

## S3 method for class 'resizingTextGrob'
preDrawDetails(x)

## S3 method for class 'resizingTextGrob'
postDrawDetails(x)

## S3 method for class 'ballGrob'
drawDetails(x, recording = TRUE)

## S3 method for class 'ballGrob'
preDrawDetails(x)

## S3 method for class 'ballGrob'
postDrawDetails(x)

## S3 method for class 'gpsim'
print(x, ...)

Arguments

object
list; pocket structure

m
width of plotting window, in nodes. By default, the nodes will be adjusted in width to fit the window.

n
length of longest genealogy

vp
viewport to draw plot in

...
arguments to be passed to textGrob.

digits
non-negative integer; number of decimal digits to print in the node time

x
An R object.

recording
A logical value indicating whether a grob is being added to the display list or redrawn from the display list.

Details

Code for the resizing text adapted from a blog post by Mark Heckmann (https://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/).

[Package phylopomp version 0.7.2.0 Index]