lbdp {phylopomp}	R Documentation

Linear birth-death-sampling model

Description

The genealogy process induced by a simple linear birth-death process with constant-rate sampling.

Usage

runLBDP(time, t0 = 0, lambda = 2, mu = 1, psi = 1, chi = 0, n0 = 5)

continueLBDP(object, time, lambda = NA, mu = NA, psi = NA, chi = NA)

lbdp_pomp(x, lambda, mu, psi, chi = 0, n0 = 1, t0 = 0)

lbdp_exact(x, lambda, mu, psi, chi = 0, n0 = 1)

Arguments

time
final time

t0
initial time

lambda
per capita birth rate

mu
per capita death rate

psi
per capita non-destructive sampling rate

chi
per capita destructive sampling rate

n0
population size at time t0

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

x
genealogy in phylopomp format (i.e., an object that inherits from ‘gpgen’).

Details

lbdp_pomp constructs a pomp object containing a given set of data and a linear birth-death-sampling process.

lbdp_exact gives the exact log likelihood of a linear birth-death process with (optionally destructive) sampling, conditioned on the population size at time 0.

Value

runLBDP and continueLBDP return objects of class ‘gpsim’ with ‘model’ attribute “LBDP”.

lbdp_exact returns the log likelihood of the genealogy. Note that the time since the most recent sample is informative.

References

A. A. King, Q. Lin, and E. L. Ionides. Exact phylodynamic likelihood via structured Markov genealogy processes. arXiv 2405.17032, 2024. doi:10.48550/arxiv.2405.17032.

A. A. King, Q. Lin, and E. L. Ionides. Markov genealogy processes. Theoretical Population Biology 143, 77–91, 2022. doi:10.1016/j.tpb.2021.11.003.

T. Stadler. Sampling-through-time in birth-death trees. Journal of Theoretical Biology 267, 396–404, 2010. doi:10.1016/j.jtbi.2010.09.010.

T. Stadler. Sampling-through-time in birth-death trees. Journal of Theoretical Biology 267, 396–404, 2010. doi:10.1016/j.jtbi.2010.09.010.

A. A. King, Q. Lin, and E. L. Ionides. Exact phylodynamic likelihood via structured Markov genealogy processes. arXiv 2405.17032, 2024. doi:10.48550/arxiv.2405.17032.

See Also

More example genealogy processes: moran, s2i2r2, seir, si2r, siir, simulate(), sir, strains, twospecies, twoundead

Examples

simulate("LBDP",time=4) |> plot(points=TRUE)

simulate("LBDP",lambda=2,mu=1,psi=3,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot()

simulate("LBDP",time=4) |>
  lineages() |>
  plot()

[Package phylopomp version 0.16.2.0 Index]