lbdp {phylopomp}	R Documentation

Linear birth-death-sampling model

Description

The genealogy process induced by a simple linear birth-death process with constant-rate sampling.

Usage

runLBDP(time, t0 = 0, lambda = 2, mu = 1, psi = 1, n0 = 5)

continueLBDP(object, time, lambda = NA, mu = NA, psi = NA)

lbdp_exact(x, lambda, mu, psi, n0 = 1)

lbdp_pomp(x, lambda, mu, psi, n0 = 1, t0 = 0)

Arguments

time
final time

t0
initial time

lambda
per capita birth rate

mu
per capita recovery rate.

psi
per capita sampling rate.

n0
initial population size

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

x
genealogy in phylopomp format (i.e., an object that inherits from ‘gpgen’).

Details

lbdp_exact gives the exact log likelihood of a linear birth-death process, conditioned on n_0 = 0 (Stadler, 2010, Thm 3.5). The derivation is also given in comments in the code.

lbdp_pomp constructs a pomp object containing a given set of data and a linear birth-death-sampling process.

Value

runLBDP and continueLBDP return objects of class ‘gpsim’ with ‘model’ attribute “LBDP”.

lbdp_exact returns the log likelihood of the genealogy. Note that the time since the most recent sample is informative.

References

T. Stadler. Sampling-through-time in birth-death trees. Journal of Theoretical Biology 267, 396–404, 2010.

See Also

More example genealogy processes: moran, seir, si2r, siir, simulate(), sir

Examples

simulate("LBDP",time=4) |> plot(points=TRUE)

simulate("LBDP",lambda=2,mu=1,psi=3,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot()

simulate("LBDP",time=4) |>
  lineages() |>
  plot()

[Package phylopomp version 0.12.0.1 Index]