lineages {phylopomp}	R Documentation

Lineage-count function

Description

The number of lineages through time

Usage

lineages(object, prune = TRUE, obscure = TRUE)

## S3 method for class 'gplin'
plot(x, ..., palette = scales::hue_pal(l = 30, h = c(220, 580)))

Arguments

object
gpsim object.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

x
object of class ‘gpsim’

...
passed to treeplot

palette
color palette for branches. This can be furnished either as a function or a vector of colors. If this is a function, it should take a single integer argument, the number of colors required. If it is a vector, ....

Value

A tibble containing the lineage count function. If the genealogy has been obscured (the default), the number in the lineages column is the total number of lineages present at the times in the time column. If the genealogy has not been obscured (obscure = FALSE), the deme-specific lineage counts are returned.

The tibble returned by lineages has a plot method.

Examples

library(tidyverse)

pal <- c("#00274c","#ffcb05")

simulate("SIIR",time=3) -> x
plot_grid(
  x |> plot(),
  x |> lineages() |> plot(),
  x |> plot(obscure=FALSE,palette=pal),
  x |> lineages(obscure=FALSE) |>
    plot(palette=pal,legend.position=c(0.8,0.9)),
  align="v",axis="b",
  ncol=2,byrow=FALSE
)

[Package phylopomp version 0.0.999.9 Index]