lineages {phylopomp}	R Documentation

Lineage-count function

Description

Lineage-counts, saturations, and event-codes.

Usage

lineages(object, prune = TRUE, obscure = TRUE)

## S3 method for class 'gplin'
plot(x, ..., palette = scales::hue_pal(l = 30, h = c(220, 580)))

Arguments

object
gpsim object.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

x
object of class ‘gpgen’

...
passed to theme.

palette
color palette for branches. This can be furnished either as a function or a vector of colors. If this is a function, it should take a single integer argument, the number of colors required. If it is a vector, it should have at least as many elements as there are demes in the genealogy.

Details

This function extracts from the specified genealogy several important time-varying quantities. These include:

lineages

number of lineages through time

saturation

the number of lineages emerging from the event

event_type

an integer coding the type of event

If the genealogy has been obscured (the default), the number in the lineages returned is the total number of lineages present at the specified time and the saturation is the total saturation. If the genealogy has not been obscured (obscure = FALSE), the deme-specific data are returned. In this case, the deme column specifies the pertinent deme.

The event types are:

0

no event,

-1

a root,

1

a sample event,

2

a non-sample event,

3

the end of the time interval, which may or may not coincide with the latest tip of the genealogy.

Value

A tibble containing information about the genealogy. See Details for specifics. The tibble returned by lineages has a plot method.

Examples

library(tidyverse)

pal <- c("#00274c","#ffcb05")

simulate("SIIR",time=3) -> x
plot_grid(
  x |> plot(),
  x |> lineages() |> plot(),
  x |> plot(obscure=FALSE,palette=pal),
  x |> lineages(obscure=FALSE) |>
    plot(palette=pal,legend.position=c(0.8,0.9)),
  align="v",axis="b",
  ncol=2,byrow=FALSE
)

[Package phylopomp version 0.10.5.0 Index]