moran {phylopomp}	R Documentation

The classical Moran model

Description

The Markov genealogy process induced by the classical Moran process, in which birth/death events occur at a constant rate and the population size remains constant.

Usage

runMoran(time, t0 = 0, mu = 1, psi = 1, n = 100)

continueMoran(object, time, mu = NA, psi = NA)

moran_exact(x, n = 100, mu = 1, psi = 1)

Arguments

time
final time

t0
initial time

mu
per capita event rate

psi
per capita sampling rate

n
population size

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

x
genealogy in phylopomp format (i.e., an object that inherits from ‘gpgen’).

Details

moran_exact gives the exact log likelihood of a genealogy under the uniformly-sampled Moran process.

Value

runMoran and continueMoran return objects of class ‘gpsim’ with ‘model’ attribute “Moran”.

moran_exact returns the log likelihood of the genealogy.

References

P.A.P. Moran. Random processes in genetics. Mathematical Proceedings of the Cambridge Philosophical Society 54, 60–71, 1958. doi:10.1017/s0305004100033193.

See Also

More example genealogy processes: lbdp, s2i2r2, seir, si2r, siir, simulate(), sir, twospecies

[Package phylopomp version 0.14.8.0 Index]