moran {phylopomp}	R Documentation

The classical Moran model

Description

The Markov genealogy process induced by the classical Moran process, in which birth/death events occur at a constant rate and the population size remains constant.

Usage

runMoran(time, t0 = 0, n = 100, mu = 1, psi = 1)

continueMoran(object, time, mu = NA, psi = NA)

Arguments

time
final time

t0
initial time

n
population size

mu
event rate

psi
sampling rate.

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

Value

An object of class ‘gpsim’ with ‘model’ attribute “Moran”.

See Also

More example genealogy processes: lbdp, seir, si2r, siir, simulate(), sir

[Package phylopomp version 0.4.0.0 Index]