newick2df {phylopomp}	R Documentation

Convert a tree in Newick format to data frame

Description

Convert a genealogical tree in Newick format to a data frame, containing event times, number of lineages, sample indicator, direct-descent indicator, and coalescent indicator.

Usage

newick2df(tree)

Arguments

tree
tree data in Newick format.

Value

A data frame suitable for use as pomp input. This data frame will have three columns:

time

numeric

lineages

integer; the lineage-count function

code

integer; a code describing the nature of the event. 0 indicates a coalescence; 1 indicates a dead sample; 2 indicates a live sample.

Examples

playSIRwS(beta=2,gamma=1,psi=2,S0=100,I0=2,times=c(0,5),t0=0,tree=TRUE) -> x
y <- getInfo(x)
newick2df(y$tree) -> z

# compare to y$lineages
y$lineages %>%
  all.equal(tail(z,-1)[,1:2],tolerance=1e-5)

[Package phylopomp version 0.0.8.2 Index]