parse_newick {phylopomp}	R Documentation

parse_newick

Description

Parse a Newick description and extract various equivalent representations.

Usage

parse_newick(
  x,
  prune = TRUE,
  obscure = TRUE,
  t0 = 0,
  time = FALSE,
  description = FALSE,
  structure = FALSE,
  yaml = FALSE,
  lineages = TRUE,
  tree = FALSE
)

Arguments

x
character; the Newick description. See Details for specifics.

prune
logical; prune the genealogy?

obscure
logical; obscure the demes?

t0
numeric; the root time.

time
logical; return the current time?

description
logical; return the description?

structure
logical; return the structure in R list format?

yaml
logical; return the structure in YAML format?

lineages
logical; return the lineage-count function?

tree
logical; return the tree?

Value

A list containing the requested elements, including any or all of:

time

the current time

description

a human readable description of the state of the genealogy process

yaml

the state of the genealogy process in YAML format

structure

the state of the genealogy process in R list format

lineages

a tibble containing the lineage count function through time

tree

the genealogical tree, in Newick format

[Package phylopomp version 0.7.2.0 Index]