parse_newick {phylopomp}	R Documentation

parse a Newick string

Description

Parses a Newick description and returns a binary version of the genealogy.

Usage

parse_newick(x, t0 = 0, tf = NA)

Arguments

x
character; the Newick description. See Details for specifics.

t0
numeric; the root time.

tf
numeric; the current or final time.

Details

parse_newick can only handle a subset of the full Newick specification. In particular, labels are assumed to be of the form <TYPE>_<DEME>_<LABEL>, i.e., each label has three parts, separated by underscores (‘_’). The parts are as follows.

TYPE must be a single character from among the following: ‘b’, ‘g’, ‘m’, ‘o’.

‘b’ signifies a sample.

‘g’ signifies an internal node.

‘m’ signifies a root.

‘o’ indicates an extant lineage.

DEME must be a non-negative integer, specifying the deme in which the branch resides. If deme information is not present, use 0.

LABEL is ignored and may be left out.

Value

An object of class “gpgen”.

[Package phylopomp version 0.15.0.0 Index]