seir {phylopomp}	R Documentation

Classical susceptible-exposed-infected-recovered model

Description

The population is structured by infection progression.

Usage

runSEIR(
  time,
  t0 = 0,
  Beta = 4,
  sigma = 1,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  E0 = 5,
  I0 = 5,
  R0 = 0
)

runSEIRS(
  time,
  t0 = 0,
  Beta = 4,
  sigma = 1,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  E0 = 5,
  I0 = 5,
  R0 = 0
)

continueSEIR(
  object,
  time,
  Beta = NA,
  sigma = NA,
  gamma = NA,
  psi = NA,
  omega = NA
)

continueSEIRS(
  object,
  time,
  Beta = NA,
  sigma = NA,
  gamma = NA,
  psi = NA,
  omega = NA
)

seirs_pomp(x, Beta, sigma, gamma, psi, omega = 0, S0, E0, I0, R0)

Arguments

time
final time

t0
initial time

Beta
transmission rate

sigma
progression rate

gamma
recovery rate

psi
per capita sampling rate

omega
rate of waning of immunity

S0, E0, I0, R0
initial sizes of S, E, I, R compartments, respectively.

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

x
genealogy in phylopomp format.

Details

seirs_pomp constructs a ‘pomp’ object containing a given set of data and an SEIRS model.

Value

runSEIR and continueSEIR return objects of class ‘gpsim’ with ‘model’ attribute “SEIR”.

seirs_pomp returns a ‘pomp’ object.

References

A. A. King, Q. Lin, and E. L. Ionides. Exact phylodynamic likelihood via structured Markov genealogy processes. arXiv 2405.17032, 2024. doi:10.48550/arxiv.2405.17032.

See Also

More example genealogy processes: lbdp, moran, s2i2r2, si2r, siir, simulate(), sir, twospecies

Examples

simulate("SEIR",Beta=2,sigma=2,gamma=1,psi=2,S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSEIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE,obscure=FALSE)

runSEIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SEIR",sigma=1,omega=1,time=20,I0=4) |> plot(obscure=FALSE)

simulate("SEIR",sigma=1,omega=1,time=20,I0=4) |>
  lineages(obscure=FALSE) |>
  plot()

[Package phylopomp version 0.14.6.1 Index]