si2r {phylopomp}	R Documentation

Two-deme model of superspreading

Description

Deme 2 consists of "superspreaders" who engender clusters of infection in "superspreading events".

Usage

runSI2R(
  time,
  t0 = 0,
  Beta = 5,
  mu = 5,
  gamma = 1,
  omega = 0,
  psi1 = 1,
  psi2 = 0,
  sigma12 = 1,
  sigma21 = 3,
  S0 = 500,
  I0 = 10,
  R0 = 0
)

continueSI2R(
  object,
  time,
  Beta = NA,
  mu = NA,
  gamma = NA,
  omega = NA,
  psi1 = NA,
  psi2 = NA,
  sigma12 = NA,
  sigma21 = NA
)

Arguments

time
final time

t0
initial time

Beta
transmission rate

mu
mean superspreading-event cluster size

gamma
recovery rate

omega
rate of waning of immunity

psi1, psi2
sampling rates for demes 1 and 2, respectively

sigma12, sigma21
movement rates from deme 1 to 2 and 2 to 1, respectively

S0
initial size of susceptible population

I0
initial size of I1 population (I2 = 0 at t = 0)

R0
initial size of recovered population

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

Details

Superspreaders (deme 2) behave differently than ordinary infections: transmission events occur at the same rate (Beta), but at each event, a superspreader infects N individuals, where

N\sim1+\mathrm{Geometric}(1/\mu).

Thus, assuming susceptibles are not limiting, the mean number of infections resulting from a superspreading event is \mu and the variance in this number is \mu^2-\mu. If susceptibles are limiting, i.e., if the number of susceptibles is not greater than N, then all remaining susceptibles are infected.

Value

runSI2R and continueSI2R return objects of class ‘gpsim’ with ‘model’ attribute “SI2R”.

See Also

More example genealogy processes: lbdp, moran, s2i2r2, seir, siir, simulate(), sir, twospecies

Examples

simulate("SI2R",time=1) |>
  plot(obscure=FALSE)

runSI2R(Beta=10,S0=2000,time=1,psi1=0) |>
  simulate(time=2,psi1=1) |>
  plot(points=TRUE,obscure=FALSE)

simulate("SI2R",time=5) |>
  lineages() |>
  plot()

simulate("SI2R",time=2) |>
  diagram(m=30)

simulate("SI2R",time=20,omega=0.2,mu=20) -> x
plot_grid(
  x |> plot(obscure=FALSE),
  x |> lineages(obscure=FALSE) |> plot(),
  ncol=1,
  align="v",axis="b"
)

[Package phylopomp version 0.14.9.1 Index]