simulate {phylopomp}	R Documentation

simulate

Description

Simulate Markov genealogy processes

Usage

simulate(object, ...)

## Default S3 method:
simulate(object, ...)

## S3 method for class 'character'
simulate(object, time, ...)

## S3 method for class 'gpsim'
simulate(object, time, ...)

Arguments

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

...
additional arguments to the model-specific simulation functions

time
end timepoint of simulation

Details

When object is of class ‘gpsim’, i.e., the result of a genealogy-process simulation, simulate acts to continue the simulation to a later timepoint. Note that, one cannot change initial conditions or t0 when continuing a simulation.

Value

An object of ‘gpsim’ class.

See Also

More example genealogy processes: lbdp, moran, seir, si2r, siir, sir

[Package phylopomp version 0.5.0.0 Index]