simulate {phylopomp}	R Documentation

simulate

Description

Simulate Markov genealogy processes

Usage

simulate(object, ...)

## Default S3 method:
simulate(object, ...)

## S3 method for class 'character'
simulate(object, time, ...)

## S3 method for class 'gpsim'
simulate(object, time, ...)

Arguments

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

...
additional arguments to the model-specific simulation functions

time
end timepoint of simulation

Details

When object is of class ‘gpsim’, i.e., the result of a genealogy-process simulation, simulate acts to continue the simulation to a later timepoint. Note that, one cannot change initial conditions or t0 when continuing a simulation.

Value

An object of ‘gpsim’ class.

References

A. A. King, Q. Lin, and E. L. Ionides. Exact phylodynamic likelihood via structured Markov genealogy processes. arXiv 2405.17032, 2024. doi:10.48550/arxiv.2405.17032.

A. A. King, Q. Lin, and E. L. Ionides. Markov genealogy processes. Theoretical Population Biology 143, 77–91, 2022. doi:10.1016/j.tpb.2021.11.003.

See Also

More example genealogy processes: lbdp, moran, s2i2r2, seir, si2r, siir, sir, strains, twospecies, twoundead

[Package phylopomp version 0.16.2.0 Index]