sir {phylopomp}	R Documentation

Classical susceptible-infected-recovered model

Description

A single, unstructured population of hosts.

Usage

runSIR(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  delta = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIR(object, time, Beta = NA, gamma = NA, psi = NA, delta = NA)

sir_pomp(data, Beta, gamma, psi, delta = 0, S0, I0, R0, t0 = 0)

Arguments

time
final time

t0
initial time

Beta
transmission rate.

gamma
recovery rate.

psi
sampling rate.

delta
immunity waning rate

S0
initial size of susceptible population.

I0
initial size of infected population.

R0
initial size of recovered population.

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

data
data frame containing the genealogy in the format returned by newick2df.

Details

sir_pomp constructs a pomp object containing a given set of data and a SIR model.

Value

An object of class ‘gpsim’ with ‘model’ attribute “SIR”.

See Also

More example genealogy processes: lbdp, moran, si2r, siir, simulate()

Examples

simulate("SIR",Beta=2,gamma=1,psi=2,S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)

runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SIR",delta=1,time=20,I0=4) |> plot()
simulate("SIR",delta=1,time=20,I0=4) |> lineages() |> plot()

[Package phylopomp version 0.0.999.9 Index]