sir {phylopomp}	R Documentation

Classical susceptible-infected-recovered model

Description

A single, unstructured population of hosts.

Usage

runSIR(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

runSIRS(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIR(object, time, Beta = NA, gamma = NA, psi = NA, omega = NA)

runSIRS(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIRS(object, time, Beta = NA, gamma = NA, psi = NA, omega = NA)

sir_pomp(x, Beta, gamma, psi, omega = 0, S0, I0, R0)

sirs_pomp(x, Beta, gamma, psi, omega = 0, S0, I0, R0)

Arguments

time
final time

t0
initial time

Beta
transmission rate.

gamma
recovery rate.

psi
sampling rate.

omega
immunity waning rate

S0, I0, R0
initial sizes of susceptible, infected, and recovered populations, respectively.

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

x
genealogy in phylopomp format (i.e., an object that inherits from ‘gpgen’).

Details

sir_pomp constructs a ‘pomp’ object containing a given set of data and a SIR model.

Value

runSIR and continueSIR return objects of class ‘gpsim’ with ‘model’ attribute “SIR”.

sir_pomp and sirs_pomp return ‘pomp’ objects.

References

A. A. King, Q. Lin, and E. L. Ionides. Exact phylodynamic likelihood via structured Markov genealogy processes. arXiv 2405.17032, 2024. doi:10.48550/arxiv.2405.17032.

A. A. King, Q. Lin, and E. L. Ionides. Markov genealogy processes. Theoretical Population Biology 143, 77–91, 2022. doi:10.1016/j.tpb.2021.11.003.

See Also

More example genealogy processes: lbdp, moran, s2i2r2, seir, si2r, siir, simulate(), strains, twospecies, twoundead

Examples

simulate("SIR",Beta=2,gamma=1,psi=2,S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)

runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SIRS",omega=1,time=20,I0=4) |> plot()
simulate("SIRS",omega=1,time=20,I0=4) |> lineages() |> plot()

[Package phylopomp version 0.16.0.0 Index]