strains {phylopomp}	R Documentation

Three strains compete for a single susceptible pool.

Description

The three demes are three distinct pathogen strains that compete for susceptibles.

Usage

runStrains(
  time,
  t0 = 0,
  Beta1 = 5,
  Beta2 = 5,
  Beta3 = 5,
  gamma = 1,
  psi1 = 1,
  psi2 = 0,
  psi3 = 0,
  S0 = 10000,
  I1_0 = 10,
  I2_0 = 10,
  I3_0 = 10,
  R0 = 0
)

continueStrains(
  object,
  time,
  Beta1 = NA,
  Beta2 = NA,
  Beta3 = NA,
  gamma = NA,
  psi1 = NA,
  psi2 = NA,
  psi3 = NA
)

strains_pomp(
  x,
  Beta1,
  Beta2,
  Beta3,
  gamma,
  psi1,
  psi2,
  psi3,
  S0,
  I1_0,
  I2_0,
  I3_0,
  R0
)

Arguments

time
final time

t0
initial time

Beta1, Beta2, Beta3
transmission rate for strains 1, 2, 3, respectively

gamma
recovery rate

psi1, psi2, psi3
sampling rates

S0
initial size of susceptible population

I1_0, I2_0, I3_0
initial numbers of strain-specific infections

R0
initial size of immune population

object
either the name of the model to simulate or a previously computed ‘gpsim’ object

x
genealogy in phylopomp format (i.e., an object that inherits from ‘gpgen’).

Details

strains_pomp constructs a ‘pomp’ object containing a given set of data and the Strains model.

Value

runStrains and continueStrains return objects of class ‘gpsim’ with ‘model’ attribute “Strains”.

strains_pomp returns a ‘pomp’ object.

See Also

More example genealogy processes: lbdp, moran, s2i2r2, seir, si2r, siir, simulate(), sir, twospecies, twoundead

[Package phylopomp version 0.16.0.0 Index]